aws-cdk/requirements.txt
aws-cdk.core
aws-cdk.aws-ec2
aws-cdk.aws-s3